Conflicts:
	applications/plugins/Flexible/types/ComponentSpecializations.cpp.inl
	applications/plugins/SofaPython/PythonEnvironment.cpp
	applications/plugins/image/CMakeLists.txt
	framework/sofa/helper/io/Image.cpp
	framework/sofa/helper/system/FileSystem.cpp
	modules/SofaBaseMechanics/SofaBaseMechanics_test/DiagonalMass_test.cpp
	modules/sofa/simulation/simulation_test/graph/Simulation_test.cpp

Former-commit-id: 8b9e5e678163aab32082dbfdee65ca88860f98f1

+++ modules/SofaBaseMechanics/MappedObject.h
@@ -86,7 +86,7 @@ public:
 
     int getSize() const
     {
-        return f_X.getValue().size();
+        return (int)f_X.getValue().size();
     }
 
     Data< VecCoord >* write(core::VecCoordId v)
